update graphql  - getNDaysRoyalty Resolver 추가  - RewardSchema 삭제  - 기타 버그 수정등
