+++ restapis-graphql/src/schema/private/UserDocumentHistory.js
@@ -40,6 +40,4 @@ schemaComposer.Mutation.addNestedFields({
     },
   },
 });
-
-
-module.exports = schemaComposer.buildSchema();
+module.exports = {}
