+++ restapis-graphql/src/schema/public/DocumentFeaturedSchema.js
@@ -40,6 +40,4 @@ function getResolverField(utc){
   //schemaComposer.Mutation.addNestedFields(fieldsByUtc.mutation);  
 })();
 
-
-
-module.exports = schemaComposer.buildSchema();
+module.exports = {}
