+++ restapis-graphql/src/schema/public/ProfileSummarySchema.js
@@ -50,4 +50,4 @@ schemaComposer.Query.addNestedFields({
 });
 
 
-module.exports = schemaComposer.buildSchema();
+module.exports = {}
