+++ restapis-graphql/src/schema/public/UserDocumentFavorite.js
@@ -51,7 +51,4 @@ const fieldsByUtc = getResolverField(utc);
 schemaComposer.Query.addNestedFields(fieldsByUtc.query);
 //schemaComposer.Mutation.addNestedFields(fieldsByUtc.mutation);
 
-
-
-
-module.exports = schemaComposer.buildSchema();
+module.exports = {}
