+++ restapis-graphql/src/schema/type/CuratorRewardType.js
@@ -1,0 +1,15 @@
+const { schemaComposer } = require('graphql-compose');
+
+  
+schemaComposer.createObjectTC({
+    name: 'CuratorReward',
+    fields: {
+        voteDate: 'Date',
+        activeDate: 'Date',
+        documentId: 'String',
+        userId: 'String',
+        voteAmount: 'String',
+        reward: 'Float'
+    }
+});
+
