+++ restapis-graphql/src/schema/type/DocumentV2Type.js
@@ -1,0 +1,13 @@
+const { schemaComposer } = require('graphql-compose');
+
+schemaComposer.createObjectTC({
+    name: 'DocumentV2',
+    fields: {
+        _id: 'String',
+        documentId: 'String',
+        accountId: 'String',
+        created: 'Date',
+        documentName: 'String',
+        author: 'UserV2'
+    }
+});
