+++ restapis-graphql/src/schema/type/UserActiveVoteAmountType.js
@@ -1,0 +1,13 @@
+const { schemaComposer } = require('graphql-compose');
+
+schemaComposer.createObjectTC({
+  name: 'UserActiveVoteAmount',
+  fields: {
+    activeDate: 'Date',
+    voteDate: 'Date',
+    documentId: 'String',
+    userId: 'String',
+    voteAmount: 'String'
+  },
+});
+
