+++ restapis-graphql/src/schema/type/UserV2Type.js
@@ -1,0 +1,11 @@
+const { schemaComposer } = require('graphql-compose');
+
+schemaComposer.createObjectTC({
+    name: 'UserV2',
+    fields: {
+        _id: 'String',
+        email: 'String',
+        name: 'String',
+        username: 'String'
+    }
+});
