+++ restapis-graphql/src/schema/type/index.js
@@ -1,0 +1,6 @@
+require('../type/CreatorRoyaltyType')
+require('../type/CuratorRewardType')
+require('../type/UserActiveVoteAmountType')
+require('../type/UserV2Type')
+require('../type/DocumentV2Type')
+
